+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Settings: new About screen showing the app's current version.
+
 ## [0.5.0] - 2026-07-19
 
 ### Added